# api/v1/signal/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/signal/service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "SignalService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/signal/perform": {
      "post": {
        "operationId": "SignalService_PerformSignalAnalysis",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/signalPerformSignalResponseAnalysis"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/signalPerformSignalRequestAnalysis"
            }
          }
        ],
        "tags": [
          "SignalService"
        ]
      }
    },
    "/v1/signal/settings/change": {
      "post": {
        "operationId": "SignalService_ChangeTradingSettings",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/signalChangeTradingSettingsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/signalChangeTradingSettingsRequest"
            }
          }
        ],
        "tags": [
          "SignalService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "signalChangeTradingSettingsRequest": {
      "type": "object",
      "properties": {
        "interval": {
          "type": "string"
        },
        "offNotify": {
          "type": "boolean"
        }
      }
    },
    "signalChangeTradingSettingsResponse": {
      "type": "object"
    },
    "signalPerformSignalRequestAnalysis": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "signalPerformSignalResponseAnalysis": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}